Cleanup file location, move headers to include direcotry, and source to src directory.
