+++ include/parser.hpp
@@ -4,8 +4,8 @@
 #include <string>
 #include <iostream>
 #include <assert.h>
-#include <include/types.hpp>
-#include <include/utils.hpp>
+#include "types.hpp"
+#include "utils.hpp"
 
 enum {
     // We are in a non-WS state
