+++ include/phrase_map.hpp
@@ -9,7 +9,7 @@
 #include <stdio.h>
 #include <assert.h>
 
-#include <include/types.hpp>
+#include "types.hpp"
 
 using namespace std;
