+++ include/utils.hpp
@@ -8,7 +8,7 @@
 #include <assert.h>
 #include <stdio.h>
 
-#include <include/types.hpp>
+#include "types.hpp"
 
 #if defined DEBUG
 #define DCERR(X) std::cerr<<X;
